// src/core/http/generated/models/admin-partners/partnerCoupon.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * PetVex Admin - Admin Partners
 * Filtered OpenAPI specification for Admin Partners.
 * OpenAPI spec version: 0.0.1
 */
import type { PartnerCouponPartner } from './partnerCouponPartner.ts';

export interface PartnerCoupon {
  id: number;
  /** @nullable */
  partner?: PartnerCouponPartner;
  partner_id: number;
  code: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  discount_type: string | null;
  /** @nullable */
  discount_value: string | null;
  /** @nullable */
  starts_at: string | null;
  /** @nullable */
  expires_at: string | null;
  /** @nullable */
  usage_limit: number | null;
  /** @nullable */
  usage_limit_per_tenant: number | null;
  is_active: boolean;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
}